converter, storage: extract disk driver cache mode and IO
Move SetDriverCacheMode, IsPreAllocated, and SetOptimalIOMode from
converter.go to storage/driver.go.
