+++ pkg/virt-launcher/virtwrap/converter/converter_test.go
@@ -34,7 +34,6 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/onsi/gomega/gstruct"
 	"github.com/onsi/gomega/types"
-	"go.uber.org/mock/gomock"
 
 	k8sv1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
@@ -4096,168 +4095,6 @@ var _ = Describe("Converter", func() {
 			),
 		)
 	})
-})
-
-var _ = Describe("Driver Cache and IO Settings", func() {
-	var ctrl *gomock.Controller
-	var mockDirectIOChecker *MockDirectIOChecker
-
-	BeforeEach(func() {
-		ctrl = gomock.NewController(GinkgoT())
-		mockDirectIOChecker = NewMockDirectIOChecker(ctrl)
-	})
-
-	expectCheckTrue := func() {
-		mockDirectIOChecker.EXPECT().CheckBlockDevice(gomock.Any()).AnyTimes().Return(true, nil)
-		mockDirectIOChecker.EXPECT().CheckFile(gomock.Any()).AnyTimes().Return(true, nil)
-	}
-
-	expectCheckFalse := func() {
-		mockDirectIOChecker.EXPECT().CheckBlockDevice(gomock.Any()).AnyTimes().Return(false, nil)
-		mockDirectIOChecker.EXPECT().CheckFile(gomock.Any()).AnyTimes().Return(false, nil)
-	}
-
-	expectCheckError := func() {
-		checkerError := fmt.Errorf("DirectIOChecker error")
-		mockDirectIOChecker.EXPECT().CheckBlockDevice(gomock.Any()).AnyTimes().Return(false, checkerError)
-		mockDirectIOChecker.EXPECT().CheckFile(gomock.Any()).AnyTimes().Return(false, checkerError)
-	}
-
-	DescribeTable("should correctly set driver cache mode", func(cache, expectedCache string, setExpectations func()) {
-		disk := &api.Disk{
-			Driver: &api.DiskDriver{
-				Cache: cache,
-			},
-			Source: api.DiskSource{
-				File: "file",
-			},
-		}
-		setExpectations()
-		err := SetDriverCacheMode(disk, mockDirectIOChecker)
-		if expectedCache == "" {
-			Expect(err).To(HaveOccurred())
-		} else {
-			Expect(err).ToNot(HaveOccurred())
-			Expect(disk.Driver.Cache).To(Equal(expectedCache))
-		}
-	},
-		Entry("detect 'none' with direct io", "", string(v1.CacheNone), expectCheckTrue),
-		Entry("detect 'writethrough' without direct io", "", string(v1.CacheWriteThrough), expectCheckFalse),
-		Entry("fallback to 'writethrough' on error", "", string(v1.CacheWriteThrough), expectCheckError),
-		Entry("keep 'none' with direct io", string(v1.CacheNone), string(v1.CacheNone), expectCheckTrue),
-		Entry("return error without direct io", string(v1.CacheNone), "", expectCheckFalse),
-		Entry("return error on error", string(v1.CacheNone), "", expectCheckError),
-		Entry("'writethrough' with direct io", string(v1.CacheWriteThrough), string(v1.CacheWriteThrough), expectCheckTrue),
-		Entry("'writethrough' without direct io", string(v1.CacheWriteThrough), string(v1.CacheWriteThrough), expectCheckFalse),
-		Entry("'writethrough' on error", string(v1.CacheWriteThrough), string(v1.CacheWriteThrough), expectCheckError),
-	)
-
-	It("should fail to set appropriate driver cache mode for a nil disk", func() {
-		Expect(SetDriverCacheMode(nil, nil)).To(MatchError("unable to set a driver cache mode, disk is nil"))
-	})
-
-	It("should check block device paths correctly", func() {
-		disk := &api.Disk{
-			Source: api.DiskSource{Dev: "/dev/vda"},
-			Driver: &api.DiskDriver{},
-		}
-		mockDirectIOChecker.EXPECT().CheckBlockDevice("/dev/vda").Return(true, nil)
-
-		Expect(SetDriverCacheMode(disk, mockDirectIOChecker)).To(Succeed())
-		Expect(disk.Driver.Cache).To(Equal(string(v1.CacheNone)))
-	})
-
-	It("should resolve datastore block dev over frontend file source", func() {
-		disk := &api.Disk{
-			Source: api.DiskSource{
-				File: "/test/overlay.qcow2",
-				DataStore: &api.DataStore{
-					Source: &api.DiskSource{Dev: "/dev/vda"},
-				},
-			},
-			Driver: &api.DiskDriver{},
-		}
-		mockDirectIOChecker.EXPECT().CheckBlockDevice("/dev/vda").Return(true, nil)
-
-		Expect(SetDriverCacheMode(disk, mockDirectIOChecker)).To(Succeed())
-	})
-	DescribeTable("should set appropriate IO modes", func(disk *api.Disk, expectedIO v1.DriverIO, isPreAllocated bool) {
-		SetOptimalIOMode(disk, func(path string) bool { return isPreAllocated })
-		Expect(disk.Driver.IO).To(Equal(expectedIO))
-	},
-		Entry("user-specified IO",
-			&api.Disk{Driver: &api.DiskDriver{IO: v1.IOThreads}},
-			v1.IOThreads, false,
-		),
-		Entry("sparse image",
-			&api.Disk{Source: api.DiskSource{File: "test.img"}, Driver: &api.DiskDriver{}},
-			v1.DriverIO(""), false,
-		),
-		Entry("pre-allocated image with O_DIRECT",
-			&api.Disk{Source: api.DiskSource{File: "test.img"}, Driver: &api.DiskDriver{Cache: string(v1.CacheNone)}},
-			v1.IONative, true,
-		),
-		Entry("pre-allocated image without O_DIRECT",
-			&api.Disk{Source: api.DiskSource{File: "test.img"}, Driver: &api.DiskDriver{Cache: string(v1.CacheWriteThrough)}},
-			v1.DriverIO(""), true,
-		),
-		Entry("block device with O_DIRECT",
-			&api.Disk{Source: api.DiskSource{Dev: "/dev/test"}, Driver: &api.DiskDriver{Cache: string(v1.CacheNone)}},
-			v1.IONative, true,
-		),
-		Entry("datastore block device with O_DIRECT",
-			&api.Disk{
-				Source: api.DiskSource{
-					File: "/test/overlay.qcow2",
-					DataStore: &api.DataStore{
-						Type:   "block",
-						Source: &api.DiskSource{Dev: "/dev/vda"},
-					},
-				},
-				Driver: &api.DiskDriver{Cache: string(v1.CacheNone)},
-			},
-			v1.IONative, false,
-		),
-		Entry("datastore block device without O_DIRECT",
-			&api.Disk{
-				Source: api.DiskSource{
-					File: "/test/overlay.qcow2",
-					DataStore: &api.DataStore{
-						Type:   "block",
-						Source: &api.DiskSource{Dev: "/dev/vda"},
-					},
-				},
-				Driver: &api.DiskDriver{Cache: string(v1.CacheWriteThrough)},
-			},
-			v1.DriverIO(""), false,
-		),
-		Entry("datastore file backend pre-allocated with O_DIRECT",
-			&api.Disk{
-				Source: api.DiskSource{
-					File: "/test/overlay.qcow2",
-					DataStore: &api.DataStore{
-						Type:   "file",
-						Source: &api.DiskSource{File: "/disks/disk.img"},
-					},
-				},
-				Driver: &api.DiskDriver{Cache: string(v1.CacheNone)},
-			},
-			v1.IONative, true,
-		),
-		Entry("datastore file backend sparse with O_DIRECT",
-			&api.Disk{
-				Source: api.DiskSource{
-					File: "/test/overlay.qcow2",
-					DataStore: &api.DataStore{
-						Type:   "file",
-						Source: &api.DiskSource{File: "/disks/disk.img"},
-					},
-				},
-				Driver: &api.DiskDriver{Cache: string(v1.CacheNone)},
-			},
-			v1.DriverIO(""), false,
-		),
-	)
 })
 
 func diskToDiskXML(arch string, disk *v1.Disk) string {
